Implement Cython hadoofus wrappers
The wrapped API aims to mimic the existing pydoofus API as closely as
possible, modulo the limits of the underlying C API. We will need to
adapt the pydoofus unit tests and pydoofus API itself to match this API.

APIs use and return "nice" python objects that implement the appropriate
python special methods. For example, directory listings are iterable,
you can slice on located blocks objects, and every object has a decent
repr(). str() for directory listing and file status objects returns
something similar to the output of 'hadoop dfs -ls'.

Datanode objects can only read/write in-memory buffers at this time (no
zero-copy files). Additionally, there may be an extra copy in the read
path because of the C buf to python bytes conversion -- ideally there's
a python API for allocating a string in advance, and then we can change
this code to read into the allocated buffer.