Build .PYC files when Cython present, else use .C
With this change you are able to install the python module without any Makefiles: `python setup.py install` from within the directory and it's good to go. If the C files aren't already there (which they should be in a distribution downloaded from PyPI) then you'll get a warning about the hadoofus.c not being present if you don't have Cython installed.

After installing the python module, libhadoofus separately needs to be findable.