Added proper support for compiling with setup.py
Previously there was an odd mix of Makefiles and a setup.py for
compiling the python wrapper for libhadoofus. I've done away with
that, and moved all the responsibility for statically compiling
the python library into setup.py. The Makefile's still exist,
however they simply delegate to python.

This is especially handy as hadoofus becomes ready to ship off
to pypi, since setuptools requires that you use the setup.py
to compile everything.