Issues #19: Implement several HDFS2 methods
getBlockLocations, create, delete. This required bringing in several
HDFS2 object parsers: BlockTokenIdentifier ("token"), LocatedBlock,
LocatedBlocks, and DatanodeInfo. LocatedBlock and LocatedBlocks were
slightly extended to add some fields new in v2.0 or v2.2 of the
protocol. Several fields in DatanodeInfo were omitted (they also exist
and we also omit them in v1 of the protocol).

A basic testcase was added to the v1, 2.0, and 2.2 unit tests that
creates a file, tests getBlockLocations() on it, and then deletes the
file.