Fix pthread mutex and cond variable leaks in 'futures' APIs
As part of this, make the hdfs_rpc_response_future ABI private.  Allow
consumers to reuse objects to avoid some of the costs of dynamic memory
allocation, but disallow inline futures for now.

More similar leaks remain, but this is a big piece of it.

Reported by:	Alex Smith <aes7mv AT virginia.edu>