Fix cemeyer/hadoofus#59 - IP instead of hostname for dn connections
Use IP addresses instead of hostname in datanode_info for datanode connections.

v1 datanode_info uses hostname for both ipaddr and hostname. The v1
implementation hadoofus was originally tested on used an IP address for
datanode_info hostname anyways, so this shouldn't be an issue.

Also fixes a memory leak in which datanode_info hostname and port strings were
not freed.