Integrate crc32c functions into single internal API
Modify the third-party crc32c source files to work with hadoofus.

An ifunc resolves the internal crc32c() function based on whether the
hardware-accelerated implementations are available for the current
system. If the ifunc attribute is not available or the preprocessor
macro NO_IFUNC is defined, a non-ifunc crc32c() definition is used.

Add unit test for internal crc32c() function.

Note: crc32c datanode support is not yet implemented.