tests/Makefile: Don't error out on missing field initializers
Clang 3.8 mistakenly warns about use of C's idiomatic zero initializer
"= { 0 };". This was fixed in later versions of clang.

Also change src/Makefile from -Wno-missing-field-initializers to
-Wno-error=missing-field-initializers, so that missing field
initializer warnings are still visible when meaningful (this does
clutter the output for clang 3.8, but use gcc or a newer version of
clang if this bothers you, or set
CFLAGS=-Werror=missing-field-initializers).