t_datanode_nb: Process reads for the next block immediately
The newly inited datanode struct must be processed prior to the next
call to hdfs_datanode_get_eventfd(), as it is an API misuse to try to
wait on a datanode event when there is no event to wait on.

Previously, in order to proceed past HDFS_DN_ST_INITED, this test
directly called hdfs_datanode_connect_nb() after iterating to the next
located block, and hackishly assumed that hdfs_datanode_connect_nb()
would return HDFS_AGAIN on its first invocation (i.e. assuming that
the first call to connect(2) for a non-blocking socket would return
with EINPROGRESS), but this is not always the case, e.g. on some
systems connecting to localhost can return a successful connection
immediately.

Now, we loop over the entire block read logic whenever we iterate to
the next located block.

TODO: This test function is extremely unwieldy now, and should be
refactored out into multiple functions.