tests: Update test_recoverLease to be more semantically correct
Previously we asserted that recoverLease returned false then proceeded
to add a block the "recovered" file. Really, one should retry
recoverLease until it returns true, which results in the file being
closed (with no one holding the lease), then the client can reopen the
file to perform whatever work still needs to be done on the file.